feat: change the format of getting key information
