+++ gcp-kms-middleware/keyinfo.json
@@ -1,6 +1,7 @@
 {
-    "keynames" : {
+    "gcpkms" : {
         "relayer1" : "projects/rock-heaven-340205/locations/global/keyRings/xpla-test/cryptoKeys/go-test-key/cryptoKeyVersions/1",
-        "relayer2" : "projects/rock-heaven-340205/locations/global/keyRings/xpla-test/cryptoKeys/go-test-key-2/cryptoKeyVersions/1"
+        "relayer2" : "projects/rock-heaven-340205/locations/global/keyRings/xpla-test/cryptoKeys/go-test-key-2/cryptoKeyVersions/1",
+        "relayer3" : "projects/rock-heaven-340205/locations/global/keyRings/xpla-test/cryptoKeys/go-test-key-3/cryptoKeyVersions/1"
     }
 }
