+++ gcp-kms-middleware/signer/signer.go
@@ -1,0 +1,6 @@
+package signer
+
+type Signer interface {
+	GetPublicKey() (string, error)
+	Sign(base64ToSign string) (string, error)
+}
